requirements.txt
flask
flask-script
flask-mongoengine
flask-wtf
py-bcrypt